Rename LazyLoadingCompletedHomeworksBloc file.
